# docs/swagger.yaml
basePath: /
definitions:
  handlers.BasicResponse:
    properties:
      id:
        type: integer
      message:
        type: string
    type: object
  handlers.CostResponse:
    properties:
      message:
        type: string
      sum_cost:
        type: integer
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handlers.ListResponse:
    properties:
      message:
        type: string
      meta:
        $ref: '#/definitions/handlers.Metadata'
      subscriptions:
        items:
          $ref: '#/definitions/subs.Subscription'
        type: array
    type: object
  handlers.Metadata:
    properties:
      limit:
        type: integer
      page:
        type: integer
      pages:
        type: integer
      total:
        type: integer
    type: object
  handlers.SubscriptionResponse:
    properties:
      message:
        type: string
      subscription:
        $ref: '#/definitions/subs.Subscription'
    type: object
  handlers.basicRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  subs.Subscription:
    properties:
      cost:
        type: integer
      endDate:
        type: string
      id:
        type: integer
      service:
        type: string
      startDate:
        type: string
      userID:
        type: string
    type: object
info:
  contact: {}
  description: API for managing user subscriptions.
  title: Subscriptions Service API
  version: "1.0"
paths:
  /subscriptions/v1/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Subscription payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.basicRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.BasicResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/v1/delete/{id}:
    delete:
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BasicResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete subscription
      tags:
      - subscriptions
  /subscriptions/v1/get/{id}:
    get:
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get subscription by ID
      tags:
      - subscriptions
  /subscriptions/v1/get/query:
    get:
      parameters:
      - description: Service name
        in: query
        name: service
        required: true
        type: string
      - description: User UUID
        in: query
        name: userID
        required: true
        type: string
      - description: Start date MM-YYYY
        in: query
        name: startDate
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get subscription by unique params
      tags:
      - subscriptions
  /subscriptions/v1/list:
    get:
      parameters:
      - description: Page
        in: query
        name: Page
        type: integer
      - description: Limit
        in: query
        name: Limit
        type: integer
      - description: Sort (cost_asc\|cost_desc\|service_asc\|service_desc\|start_date)
        in: query
        name: sort
        type: string
      - description: Service name
        in: query
        name: service
        type: string
      - description: User UUID
        in: query
        name: userID
        type: string
      - description: Start date MM-YYYY
        in: query
        name: startDate
        type: string
      - description: End date MM-YYYY
        in: query
        name: endDate
        type: string
      - description: Cost
        in: query
        name: price
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List subscriptions
      tags:
      - subscriptions
  /subscriptions/v1/total:
    get:
      parameters:
      - description: Start date MM-YYYY
        in: query
        name: startDate
        required: true
        type: string
      - description: End date MM-YYYY
        in: query
        name: endDate
        required: true
        type: string
      - description: Service name
        in: query
        name: service
        type: string
      - description: User UUID
        in: query
        name: userID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get total subscription cost for period
      tags:
      - subscriptions
  /subscriptions/v1/update/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      - description: Subscription payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.basicRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BasicResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update subscription
      tags:
      - subscriptions
produces:
- application/json
schemes:
- http
swagger: "2.0"
tags:
- name: subscriptions